oh-tab-b7qq: extract conversation store root
